fix(mailbox): premiumDaily 'operator is not unique: date + unknown' 수정
WINDOW_DAYS가 바인드 파라미터(unknown 타입)로 전달돼 date+unknown 연산자 모호(42725)로
매 레이아웃 로드마다 ensurePremiumDailyMail 실패(fire-and-forget라 페이지는 200·기능만 미동작).
+(${WINDOW_DAYS})::int 명시 캐스트로 date+int 확정. staging/prod 재현·수정 검증.

+++ lib/game/mailbox/premium-daily.ts
@@ -31,7 +31,7 @@ export async function ensurePremiumDailyMail(userId: string, serverId: number): 
         and sp.server_id = ${serverId}
         and sp.product_id = ${PREMIUM.id}
         and (now() at time zone 'Asia/Seoul')::date
-            <= (sp.updated_at at time zone 'Asia/Seoul')::date + ${WINDOW_DAYS}
+            <= (sp.updated_at at time zone 'Asia/Seoul')::date + (${WINDOW_DAYS})::int
     ),
     g as (
       insert into premium_daily_grants (user_id, server_id, kst_day)
